# internal/httpserver/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BoardOfDirectors API
  version: "1.0.0"
  description: |
    API для управления пользователями и играми с event-sourced логикой партии.
servers:
  - url: http://localhost:8000
paths:
  /auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
                name:
                  type: string
                avatar_url:
                  type: string
              required: [login, password, name]
      responses:
        "201":
          description: Registered user and JWT token
  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
              required: [login, password]
      responses:
        "200":
          description: Authenticated user and JWT token
  /auth/me:
    get:
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
  /auth/password:
    put:
      summary: Change current user password
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Password changed
  /leaderboard:
    get:
      summary: Get leaderboard
      description: Returns ranked player profiles. Use `sort=winrate`, `sort=respect`, or `sort=accuracy` to choose the primary weekly rating metric.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: period
          schema:
            type: string
            enum: [week, month, all]
            default: week
        - in: query
          name: sort
          schema:
            type: string
            enum: [winrate, respect, accuracy]
            default: winrate
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
  /users/me/profile:
    get:
      summary: Get current user profile and stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current profile
    put:
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                avatar_url:
                  type: string
                company_position:
                  type: string
                  description: Player's display position inside the fictional company. Cosmetic only.
              required: [name, avatar_url]
      responses:
        "200":
          description: Updated profile
  /users/{id}/profile:
    get:
      summary: Get user profile and stats
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: User profile, stats, respect count, and whether the viewer has respected this user
  /users/{id}/respect:
    post:
      summary: Give +respect to a user
      description: A viewer can respect another user once. Repeating the request is idempotent.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Updated user profile
  /users:
    post:
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required: [name]
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                required: [user]

    get:
      summary: List users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                required: [users]
  /users/{id}:
    get:
      summary: Get user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                required: [user]

    put:
      summary: Update user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required: [name]
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"

    delete:
      summary: Delete user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /games:
    post:
      summary: Create game
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGameRequest"
      responses:
        "201":
          description: Created game
          content:
            application/json:
              schema:
                type: object
                properties:
                  game:
                    $ref: "#/components/schemas/Game"
                  state:
                    $ref: "#/components/schemas/PublicGameState"

    get:
      summary: List games
      description: Returns room summaries. Expired lobby rooms older than one hour may be removed before the list is returned.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Games list
          content:
            application/json:
              schema:
                type: object
                properties:
                  games:
                    type: array
                    items:
                      $ref: "#/components/schemas/Game"
  /games/{id}:
    get:
      summary: Get game by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Game
          content:
            application/json:
              schema:
                type: object
                properties:
                  game:
                    $ref: "#/components/schemas/Game"

    put:
      summary: Update game
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
              required: [title]
      responses:
        "200":
          description: Updated game
          content:
            application/json:
              schema:
                type: object
                properties:
                  game:
                    $ref: "#/components/schemas/Game"

    delete:
      summary: Delete game
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Deleted
  /games/{id}/state:
    get:
      summary: Get personalized game state
      description: Returns viewer-specific state and advances due automatic timer events, including bot turns and inactive-player replacement.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Personalized game state
          content:
            application/json:
              schema:
                type: object
                properties:
                  state:
                    $ref: "#/components/schemas/PublicGameState"
  /games/ws:
    get:
      summary: Open live games list WebSocket
      description: |
        Authenticated WebSocket endpoint. Pass the JWT as `?token=...`.
        The server sends `{ "type": "games", "games": Game[] }` messages with the current lobby list after game list changes.
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        "101":
          description: WebSocket protocol switch
  /games/{id}/ws:
    get:
      summary: Open live game state WebSocket
      description: |
        Authenticated WebSocket endpoint. Pass the JWT as `?token=...`.
        The server sends personalized `{ "type": "state", "state": PublicGameState }` messages after game actions.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        "101":
          description: WebSocket protocol switch
  /games/{id}/actions:
    post:
      summary: Send player action
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameActionRequest"
      responses:
        "200":
          description: Accepted action and updated state
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    description: Public emitted events only. Private events such as mole selection, mole objectives, and submitted votes are omitted.
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
                  state:
                    $ref: "#/components/schemas/PublicGameState"
                  game_deleted:
                    type: boolean
                    description: True when the action removed an empty lobby.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        name:
          type: string
        avatar_url:
          type: string
        company_position:
          type: string
        respect_count:
          type: integer
        respected_by_me:
          type: boolean
        last_seen_at:
          type: string
          nullable: true
        created_at:
          type: string
        updated_at:
          type: string
          nullable: true
      required: [id, name, created_at]
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        user:
          $ref: "#/components/schemas/Profile"
        games:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winrate:
          type: number
          format: double
        rating_points:
          type: integer
        respect_delta:
          type: integer
        accuracy_bps:
          type: integer
        xp:
          type: integer
        rank_title:
          type: string
      required: [rank, user, games, wins, losses, winrate, rating_points, respect_delta, accuracy_bps, xp, rank_title]
    LeaderboardResponse:
      type: object
      properties:
        period:
          type: string
          enum: [week, month, all]
        sort:
          type: string
          enum: [winrate, respect, accuracy]
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
      required: [period, sort, entries]
    Game:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        company_name:
          type: string
        company_situation:
          type: string
        created_at:
          type: string
        status:
          type: string
          enum: [lobby, started, finished]
        phase:
          type: string
          enum: [mole_objective_selection, major_voting, governance_proposal, governance_voting]
        winner:
          type: string
        current_round:
          type: integer
        player_count:
          type: integer
        player_user_ids:
          type: array
          items:
            type: integer
            format: int64
        players:
          type: array
          items:
            $ref: "#/components/schemas/GameListPlayer"
        is_member:
          type: boolean
      required: [id, title, created_at]
    GameListPlayer:
      type: object
      properties:
        user_id:
          type: integer
          format: int64
        name:
          type: string
        avatar_url:
          type: string
        company_position:
          type: string
        is_host:
          type: boolean
        is_ceo:
          type: boolean
        is_bot:
          type: boolean
      required: [user_id, name, is_host, is_ceo, is_bot]
    CreateGameRequest:
      type: object
      description: Creates a room and generates a persistent company scenario for the game.
      properties:
        title:
          type: string
      required: [title]
    Event:
      type: object
      properties:
        id:
          type: integer
          format: int64
        game_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
          nullable: true
        actor_name:
          type: string
          nullable: true
        event_type:
          type: string
        event_value:
          type: string
          nullable: true
        created_at:
          type: string
      required: [id, game_id, event_type, created_at]
    GameActionRequest:
      type: object
      properties:
        type:
          type: string
          enum:
            - join_game
            - leave_game
            - kick_player
            - ban_player
            - add_bot
            - send_chat_message
            - react_chat_message
            - start_game
            - choose_memorandum
            - select_mole_objectives
            - vote
            - submit_governance_proposal
            - skip_governance_proposal
        payload:
          oneOf:
            - $ref: "#/components/schemas/SendChatMessagePayload"
            - $ref: "#/components/schemas/ChooseMemorandumPayload"
            - $ref: "#/components/schemas/SelectMoleObjectivesPayload"
            - $ref: "#/components/schemas/VotePayload"
            - $ref: "#/components/schemas/SubmitGovernanceProposalPayload"
            - $ref: "#/components/schemas/ReactChatMessagePayload"
            - $ref: "#/components/schemas/AddBotPayload"
            - type: object
              additionalProperties: true
      required: [type]
    SendChatMessagePayload:
      type: object
      description: Chat message payload. Messages starting with `/me ` are posted as official statements from the player's effective position.
      properties:
        message:
          type: string
          maxLength: 500
      required: [message]
...